# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/gamification_challenges:
    get:
      tags:
        - GamificationChallenges
      summary: List GamificationChallenges
      description: Get a paginated list of GamificationChallenges with optional filtering and sorting
      operationId: listGamificationChallenges
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationChallengeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - GamificationChallenges
      summary: Create GamificationChallenge
      description: Create a new GamificationChallenge
      operationId: createGamificationChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGamificationChallengeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationChallenge'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    GamificationChallenge:
      type: object
      required:
        - id
        - name
        - state
        - period
        - visibility_mode
        - reward_failure
        - reward_realtime
        - report_frequency
        - user_ids
        - include_all_users
        - include_badge_ids
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        state:
          type: string
        manager_user_id:
          nullable: true
          type: string
          format: uuid
        period:
          type: string
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        visibility_mode:
          type: string
        reward_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_first_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_second_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_third_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_failure:
          type: boolean
        reward_realtime:
          type: boolean
        report_frequency:
          type: string
        last_report_date:
          nullable: true
          type: string
          format: date
        next_report_date:
          nullable: true
          type: string
          format: date
        remind_update_delay:
          nullable: true
          type: integer
          format: int32
        user_ids:
          type: object
        include_all_users:
          type: boolean
        include_badge_ids:
          type: object
        metadata:
          type: object
...
    CreateGamificationChallengeRequest:
      type: object
      required:
        - name
        - state
        - period
        - visibility_mode
        - reward_failure
        - reward_realtime
        - report_frequency
        - user_ids
        - include_all_users
        - include_badge_ids
        - metadata
      properties:
        name:
          type: string
          maxLength: 200
        state:
          type: string
        manager_user_id:
          nullable: true
          type: string
          format: uuid
        period:
          type: string
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        visibility_mode:
          type: string
        reward_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_first_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_second_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_third_badge_id:
          nullable: true
          type: string
          format: uuid
        reward_failure:
          type: boolean
        reward_realtime:
          type: boolean
        report_frequency:
          type: string
        last_report_date:
          nullable: true
          type: string
          format: date
        next_report_date:
          nullable: true
          type: string
          format: date
        remind_update_delay:
          nullable: true
          type: integer
          format: int32
        user_ids:
          type: object
        include_all_users:
          type: boolean
        include_badge_ids:
          type: object
        metadata:
          type: object
...
    GamificationChallengeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GamificationChallenge'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []