# src/presentation/http/gamification_challenge_handler.rs
//! GamificationChallenge REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{NaiveDate};

// Backbone framework imports
use backbone_core::http::{ApiResponse, BackboneCrudHandler};

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{GamificationChallengeService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateGamificationChallengeDto, UpdateGamificationChallengeDto, PatchGamificationChallengeDto, GamificationChallengeResponseDto};

use crate::domain::state_machine::{challenge_stateState, challenge_stateStateMachine, challenge_stateTransition};

/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum GamificationChallengeError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for GamificationChallengeError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for GamificationChallengeError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "GAMIFICATIONCHALLENGE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "GAMIFICATIONCHALLENGE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "GAMIFICATIONCHALLENGE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "GAMIFICATIONCHALLENGE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for GamificationChallenge.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /gamification_challenges | List with pagination |
/// | POST | /gamification_challenges | Create new |
/// | GET | /gamification_challenges/:id | Get by ID |
/// | PUT | /gamification_challenges/:id | Full update |
/// | PATCH | /gamification_challenges/:id | Partial update |
/// | DELETE | /gamification_challenges/:id | Soft delete |
/// | POST | /gamification_challenges/bulk | Bulk create |
/// | POST | /gamification_challenges/upsert | Upsert |
/// | GET | /gamification_challenges/trash | List deleted |
/// | POST | /gamification_challenges/:id/restore | Restore |
/// | DELETE | /gamification_challenges/empty | Empty trash |
/// | GET | /gamification_challenges/:id/deleted | Get deleted by ID |
/// | DELETE | /gamification_challenges/trash/:id | Permanent delete from trash |
/// | GET | /gamification_challenges/count | Count active entities |
/// | GET | /gamification_challenges/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(GamificationChallengeService::with_repository(repository));
/// let router = create_gamification_challenge_routes(service);
/// ```
pub fn create_gamification_challenge_routes(service: Arc<GamificationChallengeService>) -> Router {
    BackboneCrudHandler::<GamificationChallengeService, GamificationChallenge, CreateGamificationChallengeDto, UpdateGamificationChallengeDto, GamificationChallengeResponseDto>::routes(
        service,
        "/gamification_challenges",
    )
}

/// Create Axum router with only the read (GET) endpoints for GamificationChallenge.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_gamification_challenge_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_gamification_challenge_read_routes(service: Arc<GamificationChallengeService>) -> Router {
    BackboneCrudHandler::<GamificationChallengeService, GamificationChallenge, CreateGamificationChallengeDto, UpdateGamificationChallengeDto, GamificationChallengeResponseDto>::read_routes(
        service,
        "/gamification_challenges",
    )
}

/// Create Axum router with only the write (mutation) endpoints for GamificationChallenge.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_gamification_challenge_write_routes(service: Arc<GamificationChallengeService>) -> Router {
    BackboneCrudHandler::<GamificationChallengeService, GamificationChallenge, CreateGamificationChallengeDto, UpdateGamificationChallengeDto, GamificationChallengeResponseDto>::write_routes(
        service,
        "/gamification_challenges",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_gamification_challenge_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<GamificationChallengeService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_gamification_challenge_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}

// =============================================================================
// State Transition Handlers
// =============================================================================

/// Execute start transition on a GamificationChallenge.
///
/// POST /gamification_challenges/:id/transitions/start
pub async fn start_transition(
    axum::extract::State(service): axum::extract::State<Arc<GamificationChallengeService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<GamificationChallengeResponseDto>::not_found("GamificationChallenge", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<GamificationChallengeResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = challenge_stateTransition::Start.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "gamification_challenge:transition:start");
        let has_update_perm = auth.permissions.iter().any(|p| p == "gamification_challenge:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<GamificationChallengeResponseDto>::error("Insufficient permissions for start transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: challenge_stateState = entity.state.to_string().parse()
        .unwrap_or(challenge_stateState::default());
    let sm = challenge_stateStateMachine::from_state(current_state);
    if !sm.can_transition(challenge_stateTransition::Start) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<GamificationChallengeResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Inprogress".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: GamificationChallengeResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<GamificationChallengeResponseDto>::not_found("GamificationChallenge", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<GamificationChallengeResponseDto>::error(e.to_string()))),
    }
}

/// Execute close transition on a GamificationChallenge.
///
/// POST /gamification_challenges/:id/transitions/close
pub async fn close_transition(
    axum::extract::State(service): axum::extract::State<Arc<GamificationChallengeService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<GamificationChallengeResponseDto>::not_found("GamificationChallenge", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<GamificationChallengeResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = challenge_stateTransition::Close.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "gamification_challenge:transition:close");
        let has_update_perm = auth.permissions.iter().any(|p| p == "gamification_challenge:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<GamificationChallengeResponseDto>::error("Insufficient permissions for close transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: challenge_stateState = entity.state.to_string().parse()
        .unwrap_or(challenge_stateState::default());
    let sm = challenge_stateStateMachine::from_state(current_state);
    if !sm.can_transition(challenge_stateTransition::Close) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<GamificationChallengeResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Done".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: GamificationChallengeResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<GamificationChallengeResponseDto>::not_found("GamificationChallenge", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<GamificationChallengeResponseDto>::error(e.to_string()))),
    }
}

/// Execute reset transition on a GamificationChallenge.
///
/// POST /gamification_challenges/:id/transitions/reset
pub async fn reset_transition(
    axum::extract::State(service): axum::extract::State<Arc<GamificationChallengeService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<GamificationChallengeResponseDto>::not_found("GamificationChallenge", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<GamificationChallengeResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = challenge_stateTransition::Reset.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "gamification_challenge:transition:reset");
        let has_update_perm = auth.permissions.iter().any(|p| p == "gamification_challenge:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<GamificationChallengeResponseDto>::error("Insufficient permissions for reset transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: challenge_stateState = entity.state.to_string().parse()
        .unwrap_or(challenge_stateState::default());
    let sm = challenge_stateStateMachine::from_state(current_state);
    if !sm.can_transition(challenge_stateTransition::Reset) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<GamificationChallengeResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Draft".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: GamificationChallengeResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<GamificationChallengeResponseDto>::not_found("GamificationChallenge", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<GamificationChallengeResponseDto>::error(e.to_string()))),
    }
}

/// Create routes for state transitions.
pub fn create_gamification_challenge_transition_routes(service: Arc<GamificationChallengeService>) -> Router {
    use axum::routing::post;

    Router::new()
        .route("/gamification_challenges/:id/transitions/start", post(start_transition))
        .route("/gamification_challenges/:id/transitions/close", post(close_transition))
        .route("/gamification_challenges/:id/transitions/reset", post(reset_transition))
        .with_state(service)
}